still playing with designing the rest apis. For the purposes of this project, I think I like where they are in this commit. make, model, year could be more of a filter tye rest operation in the future.

+++ server/src/test/java/com/maintenance/vehicle/VehicleTestUtils.java
@@ -10,7 +10,7 @@ public class VehicleTestUtils {
      * Utility for creating a mock vehicle. Since vehicle is immutable we don't put setters on it. Instead of adding
      * setters just for unit test we use hacky reflection
      */
-    public static Vehicle getMockVehicle(int id, String make, String model, int year) throws Exception {
+    public static Vehicle getMockVehicle(int id, String make, String model, short year) throws Exception {
         Vehicle vehicleMock = new Vehicle();
         Field idField = Vehicle.class.getSuperclass().getDeclaredField("id");
         idField.setAccessible(true);
